decomp: retail-evidence passes on the cse-2.95.2 game compiler
Second run of scripts/retail-evidence-passes.py (cc1 4dfa3cf0): one
extern-width and one gp .extern win in pending bodies (fun_0022f5b0,
fun_0022eaa8). NON_MATCHING bodies only; INCLUDE_ASM stays authoritative.
./verify-baseline.sh: PASS (SHA-256 e0505810...c0c9).

+++ src/assembly/textbin/fun_0022eaa8.c
@@ -44,7 +44,7 @@ extern s32 D_0015ED5C;
 extern s32 D_0015ED84;
 extern s32 D_0015EE48;
 extern s32 D_0015EE4A;
-extern s32 D_0015F43C;
+extern f32 D_0015F43C;
 extern s32 D_0015F618;
 extern s32 D_00160404;
 extern u8 D_00160460[];
